Updates the ArePeopleWearing web project to start a single page app
This adds the Hot Towel Angular package and removes some of the folders
that are no longer necessary for the project. It fixes at least one
package reference that had gone bad for the web project. It also fixes a
bug with the RouteConfig class not being called on app_start.